+++ index.html
@@ -4,8 +4,9 @@
         <meta charset="UTF-8" />
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <link rel="stylesheet" href="styles.css" />
-        <script src="script.js" defer></script>
+        <link rel="stylesheet" href="css/index.css" />
+        <script type="module" src="js/index.js" defer async></script>
+
         <title>Document</title>
     </head>
     <body>
